Extract chat input
